+++ apps/docs/src/app/platform/component-docs/platform-search-field/platform-search-field-examples/platform-search-field-basic-example.component.ts
@@ -8,55 +8,54 @@ import { SuggestionItem, SearchInput } from '@fundamental-ngx/platform/search-fi
     styleUrls: ['./platform-search-field-basic-example.component.scss'],
     encapsulation: ViewEncapsulation.None
 })
-export class PlatformSearchFieldBasicExampleComponent implements OnInit {
-    public suggestions: SuggestionItem[];
+export class PlatformSearchFieldBasicExampleComponent {
+    suggestions: SuggestionItem[] = [
+        {
+            value: 'Apple'
+        },
+        {
+            value: 'Banana'
+        },
+        {
+            value: 'Blueberry'
+        },
+        {
+            value: 'Cherry'
+        },
+        {
+            value: 'Grape'
+        },
+        {
+            value: 'Lemon'
+        },
+        {
+            value: 'Lime'
+        },
+        {
+            value: 'Orange'
+        },
+        {
+            value: 'Peach'
+        },
+        {
+            value: 'Pineapple'
+        },
+        {
+            value: 'Plum'
+        },
+        {
+            value: 'Raspberry'
+        }
+    ];
 
-    public searchTerm = '';
-    public inputText = '';
+    searchTerm = '';
+    inputText = '';
 
-    public compactSearchTerm = '';
-    public compactInputText = '';
+    compactSearchTerm = '';
+    compactInputText = '';
 
-    ngOnInit(): void {
-        this.suggestions = [
-            {
-                value: 'Apple'
-            },
-            {
-                value: 'Banana'
-            },
-            {
-                value: 'Blueberry'
-            },
-            {
-                value: 'Cherry'
-            },
-            {
-                value: 'Grape'
-            },
-            {
-                value: 'Lemon'
-            },
-            {
-                value: 'Lime'
-            },
-            {
-                value: 'Orange'
-            },
-            {
-                value: 'Peach'
-            },
-            {
-                value: 'Pineapple'
-            },
-            {
-                value: 'Plum'
-            },
-            {
-                value: 'Raspberry'
-            }
-        ];
-    }
+    searchWithoutSuggestionsTerm = '';
+    inputWithoutSuggestionsText = '';
 
     onSearchSubmit($event: SearchInput): void {
         this.searchTerm = $event.text;
@@ -72,5 +71,13 @@ export class PlatformSearchFieldBasicExampleComponent implements OnInit {
 
     onCompactInputChange($event: SearchInput): void {
         this.compactInputText = $event.text;
+    }
+
+    onSearchWithoutSuggestionsSubmit($event: SearchInput): void {
+        this.searchWithoutSuggestionsTerm = $event.text;
+    }
+
+    onInputWithoutSuggestionsChange($event: SearchInput): void {
+        this.inputWithoutSuggestionsText = $event.text;
     }
 }