+++ e2e/wdio/platform/fixtures/appData/approval-flow-contents.ts
@@ -6,3 +6,11 @@ export const node_statuses = ['approved, in progress, not started'];
 export const remainder_text = 'Reminder has been sent to ';
 export const approved_node_status = 'approved';
 export const rejected_node_status = 'rejected';
+export const nodeOptionsArr = [
+    'Add approvers before',
+    'Add approvers after',
+    'Add parallel approvers',
+    'Edit approver',
+    'Remove'
+];
+export const dueDateMessage = 'Due in';