+++ libs/platform/src/lib/form/combobox/combobox.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { OverlayModule } from '@angular/cdk/overlay';
 
 import { ButtonModule } from '@fundamental-ngx/core/button';
 import { DynamicComponentService, TemplateModule } from '@fundamental-ngx/core/utils';
 import { InputGroupModule } from '@fundamental-ngx/core/input-group';
 import { ListModule } from '@fundamental-ngx/core/list';
 import { FormModule } from '@fundamental-ngx/core/form';
+import { PopoverModule } from '@fundamental-ngx/core/popover';
+
 import { ComboboxComponent } from './combobox/combobox.component';
 import { ComboboxHighlightPipe } from './pipes/highlight.pipe';
 import { PlatformAutoCompleteModule } from '../auto-complete/auto-complete.module';
@@ -20,7 +21,7 @@ import { PlatformAutoCompleteModule } from '../auto-complete/auto-complete.modul
         InputGroupModule,
         ListModule,
         ButtonModule,
-        OverlayModule,
+        PopoverModule,
         PlatformAutoCompleteModule,
         FormModule
     ],