+++ libs/platform/src/lib/form/multi-input/base-multi-input.ts
@@ -137,10 +137,6 @@ export abstract class BaseMultiInput extends CollectionBaseInput implements Afte
     }
 
     set value(value: any) {
-        if (!value) {
-            return;
-        }
-
         const selectedItems = Array.isArray(value) ? value : [value];
         this.setAsSelected(this._convertToOptionItems(selectedItems));
         super.setValue(value);
@@ -337,18 +333,16 @@ export abstract class BaseMultiInput extends CollectionBaseInput implements Afte
 
     /** write value for ControlValueAccessor */
     writeValue(value: any): void {
-        if (!value) {
-            return;
-        }
-
         const selectedItems = Array.isArray(value) ? value : [value];
         this.setAsSelected(this._convertToOptionItems(selectedItems));
         super.writeValue(value);
     }
 
     /** @hidden */
-    popoverOpenChangeHandle(isOpen: boolean): void {
-        this.isOpen ? this.close() : this.open();
+    _popoverOpenChangeHandle(isOpen: boolean): void {
+        if (!isOpen) {
+            this.close();
+        }
     }
 
     /** Opens the select popover body. */